+++ assets/Assets.php
@@ -10,6 +10,8 @@
 namespace humhub\modules\tasks\assets;
 
 use humhub\components\assets\AssetBundle;
+use humhub\modules\tasks\controllers\ListController;
+use humhub\modules\tasks\controllers\SearchController;
 
 class Assets extends AssetBundle
 {
@@ -37,4 +39,19 @@ class Assets extends AssetBundle
         'js/humhub.task.checklist.js',
         'js/humhub.task.search.js',
     ];
+
+    /**
+     * @inheritdoc
+     */
+    public static function register($view)
+    {
+        if ($view->context instanceof ListController || $view->context instanceof SearchController) {
+            $colorLink = $view->theme->variable('link');
+            $colorInfo = $view->theme->variable('info');
+            $view->registerCss('#task-space-menu li.active a{color:' . $colorLink . '}'
+                . '.task-overview #task-filter-nav .task-bottom-panel .filterInput[data-filter-type=checkbox] .fa.fa-check-square-o{border-color:' . $colorInfo . ';background:' . $colorInfo . '}');
+        }
+
+        return parent::register($view);
+    }
 }